agrego los demas screens
