+++ app/src/main/java/edu/unicauca/aplimovil/taskzen/MainActivity.kt
@@ -18,38 +18,12 @@ class MainActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContent {
-
-            TaskZenTheme {
-                // A surface container using the 'background' color from the theme
-                Surface(
-                    modifier = Modifier.fillMaxSize(),
-                    color = MaterialTheme.colorScheme.background
-                ) {
-                    Greeting("Android")
-                }
-            }
-
             MyApp()
-
         }
     }
 }
 
 @Composable
-
-fun Greeting(name: String, modifier: Modifier = Modifier) {
-    Text(
-        text = "Hello $name!",
-        modifier = modifier
-    )
-}
-
-@Preview(showBackground = true)
-@Composable
-fun GreetingPreview() {
-    LoginScreenPreview()
-}
-
 fun MyApp() {
     val navController = rememberNavController()
     MaterialTheme {
