+++ app/src/main/java/edu/unicauca/aplimovil/taskzen/ui/Login_Register/RegisterScreen.kt
@@ -1,0 +1,4 @@
+package edu.unicauca.aplimovil.taskzen.ui.Login_Register
+
+class RegisterScreen {
+}
